Flytter valideringen til vilkårsvurderingen fordi det er bedre at SB får beskjed her enn ved sending til beslutter. Denne valideringen kjører dessuten også når man sender til beslutter.
