+++ src/test/kotlin/no/nav/familie/ba/sak/beregning/UtbetalingssikkerhetTest.kt
@@ -176,85 +176,6 @@ class UtbetalingssikkerhetTest {
     }
 
     @Test
-    fun `Skal kaste feil når et barn har perioder utover 0-18 år`() {
-        val søker = tilfeldigPerson(personType = PersonType.SØKER)
-        val barn = tilfeldigPerson()
-        val personopplysningGrunnlag = PersonopplysningGrunnlag(
-                behandlingId = 1,
-                personer = mutableSetOf(søker, barn)
-        )
-
-        val tilkjentYtelse = lagInitiellTilkjentYtelse()
-
-        tilkjentYtelse.andelerTilkjentYtelse.addAll(listOf(
-                lagAndelTilkjentYtelse(barn.fødselsdato.minusYears(1).toString(),
-                                       barn.fødselsdato.plusYears(10).toString(),
-                                       YtelseType.UTVIDET_BARNETRYGD,
-                                       1054,
-                                       person = barn)
-        ))
-
-        val feil = assertThrows<UtbetalingsikkerhetFeil> {
-            TilkjentYtelseValidering.validerAtTilkjentYtelseKunHarGyldigTotalPeriode(tilkjentYtelse,
-                                                                                     personopplysningGrunnlag)
-        }
-
-        assertTrue(feil.frontendFeilmelding?.contains("${barn.personIdent.ident} har utbetalinger utover 0-18 år")!!)
-    }
-
-    @Test
-    fun `Skal kaste feil når søker har perioder utover 0-18 år`() {
-        val søker = tilfeldigPerson(personType = PersonType.SØKER)
-        val barn = tilfeldigPerson()
-        val personopplysningGrunnlag = PersonopplysningGrunnlag(
-                behandlingId = 1,
-                personer = mutableSetOf(søker, barn)
-        )
-
-        val tilkjentYtelse = lagInitiellTilkjentYtelse()
-
-        tilkjentYtelse.andelerTilkjentYtelse.addAll(listOf(
-                lagAndelTilkjentYtelse(barn.fødselsdato.minusYears(1).toString(),
-                                       barn.fødselsdato.plusYears(10).toString(),
-                                       YtelseType.UTVIDET_BARNETRYGD,
-                                       1054,
-                                       person = søker)
-        ))
-
-        val feil = assertThrows<UtbetalingsikkerhetFeil> {
-            TilkjentYtelseValidering.validerAtTilkjentYtelseKunHarGyldigTotalPeriode(tilkjentYtelse,
-                                                                                     personopplysningGrunnlag)
-        }
-
-        assertTrue(feil.frontendFeilmelding?.contains("Søker har utbetalinger")!!)
-    }
-
-    @Test
-    fun `Skal ikke kaste feil når barn har riktig 0-18 år periode`() {
-        val søker = tilfeldigPerson(personType = PersonType.SØKER)
-        val barn = tilfeldigPerson()
-        val personopplysningGrunnlag = PersonopplysningGrunnlag(
-                behandlingId = 1,
-                personer = mutableSetOf(søker, barn)
-        )
-
-        val tilkjentYtelse = lagInitiellTilkjentYtelse()
-
-        tilkjentYtelse.andelerTilkjentYtelse.addAll(listOf(
-                lagAndelTilkjentYtelse(barn.fødselsdato.førsteDagINesteMåned().toString(),
-                                       barn.fødselsdato.plusYears(18).sisteDagIForrigeMåned().toString(),
-                                       YtelseType.UTVIDET_BARNETRYGD,
-                                       1054,
-                                       person = barn)
-        ))
-
-        assertDoesNotThrow {
-            TilkjentYtelseValidering.validerAtTilkjentYtelseKunHarGyldigTotalPeriode(tilkjentYtelse,
-                                                                                     personopplysningGrunnlag)
-        }
-    }
-
-    @Test
     fun `Skal kaste feil når barn får har over 100% gradering for ytelsetype`() {
         val barn = tilfeldigPerson()
         val tilkjentYtelse = lagInitiellTilkjentYtelse()
